requirements.txt
aiogram
aiogram-dialog
sqlalchemy
asyncpg
apscheduler
aiocryptopay
aioplatega
aiohttp
greenlet
fastapi
uvicorn
ormsgpack
openpyxl
nats-py
pydantic
pytz
environs
redis
